requirements.txt
email-validator==2.1.0
flask==3.0.2
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
langchain==0.1.12
langchain-community==0.0.29
langchain-core==0.1.36
langchain-google-genai==0.0.10
langchain-groq==0.0.1
langgraph==0.0.55
matplotlib==3.8.3
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
plotly==5.19.0
psycopg2-binary==2.9.9
pydantic==2.6.3
python-dotenv==1.0.1
rdflib==7.0.0
requests-toolbelt==1.0.0
scipy==1.12.0
seaborn==0.13.1